// src/app/main/dashboard/constant/dashboard.enums.ts
export enum DASHBOARD {
    COMPANY = 'api/referencedata/all/company',
    GENDER = 'api/referencedata/all/gender',    
    BRAND_TYPE = 'api/referencedata/all/brandType',
    CATEGORY = 'api/referencedata/all/category',
    SUB_CATEGORY = 'api/referencedata/all/subCategory',
    AG = 'api/referencedata/all/ag',
    BRAND_NAME = 'api/referencedata/all/brandName',
    PERIOD = 'api/referencedata/all/period'
}

export enum DashboardPath {
    DISTRIBUTIONAPIPATH = 'api/limits/distribution',
    HEADERS = 'api/limits/header/distribution',
    DOWNLOAD = 'api/limits/download/distribution'
}